+++ app/javascript/react/components/MarkdownEditor/Buttons/index.js
@@ -1,0 +1,5 @@
+export {icons, labels} from './Details';
+export {default as Table} from './Table';
+export {default as Heading} from './Heading';
+export {default as LinkMenu} from './LinkMenu';
+export {default as List} from './List';
