+++ app/javascript/react/components/MarkdownEditor/milkdown_editor.css
@@ -98,7 +98,7 @@
 
 .markdown_editor .md_editor-buttons .spacer {
   align-self: center;
-  padding: 0 1rem;
+  padding: 0 .5ch;
 }
 
 .markdown_editor .md_editor-buttons .spacer:before {
